requirements.txt
click==8.0.3
flake8==4.0.1
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.1
Werkzeug==2.0.2
